Refactoring bench 2842279
